Resolving changes
