+++ test/types/local_request_body_test.exs
@@ -5,7 +5,7 @@ defmodule Kadena.Types.LocalRequestBodyTest do
 
   use ExUnit.Case
 
-  alias Kadena.Chainweb.JSON
+  alias Kadena.Chainweb.Pact.JSONRequestBody
 
   alias Kadena.Types.{
     CommandPayloadStringifiedJSON,
@@ -68,7 +68,7 @@ defmodule Kadena.Types.LocalRequestBodyTest do
       json_result: json_result
     } do
       local_request_body = LocalRequestBody.new(hash: hash, sigs: sigs, cmd: cmd)
-      ^json_result = JSON.parse(local_request_body)
+      ^json_result = JSONRequestBody.parse(local_request_body)
     end
   end
 end
